+++ public/js/reports/reports.module.js
@@ -1,5 +1,5 @@
 (function () {
     'use strict';
 
-    angular.module('app.reports', ['ngRoute', 'gettext', 'app.core']);
+    angular.module('app.reports', ['ngRoute', 'gettext', 'ui.bootstrap', 'app.core']);
 })();