+++ public/js/reports/reports.service.js
@@ -3,15 +3,27 @@
 
     angular.module('app.reports').factory('reportsService', reportsService);
 
-    reportsService.$inject = [];
+    reportsService.$inject = ['gettext', 'gettextCatalog'];
 
-    function reportsService () {
+    function reportsService (gettext, gettextCatalog) {
         let storedReport = {};
+        const aggregations = {
+            sum: gettext('Sum'),
+            avg: gettext('Avg'),
+            min: gettext('Min'),
+            max: gettext('Max'),
+            count: gettext('Count'),
+            countDistinct: gettext('Count distinct'),
+        };
+
         const service = {
             generateQuery: generateQuery,
             getQueryForFilter: getQueryForFilter,
+            getAggregationDescription: getAggregationDescription,
             storeReport: storeReport,
             getStoredReport: getStoredReport,
+            getColumnId: getColumnId,
+            getColumnDescription: getColumnDescription,
         };
 
         return service;
@@ -72,6 +84,19 @@
             return query;
         }
 
+        /**
+         * Returns a human-readable description of an aggregation function
+         *
+         * @param {string} aggregation - Name of the aggregation function
+         * @returns {string} The translated human-readable description of the
+         *     aggregation function
+         */
+        function getAggregationDescription (aggregation) {
+            if (aggregation in aggregations) {
+                return gettextCatalog.getString(aggregations[aggregation]);
+            }
+        }
+
         function storeReport (report) {
             storedReport = report;
         }
@@ -79,6 +104,36 @@
         function getStoredReport () {
             return storedReport;
         }
+
+        /*
+         * The id of a column (column.id) differs from the id of the element
+         * which that column uses (column.elementID). This allows for multiple
+         * columns which use the same element, for example to use different
+         * aggregations
+         */
+        function getColumnId (element) {
+            var columnId;
+
+            var aggregation = element.aggregation || element.defaultAggregation;
+
+            if (!aggregation) {
+                columnId = 'e' + element.elementID.toLowerCase() + 'raw';
+            } else {
+                columnId = 'e' + element.elementID.toLowerCase() + aggregation.substring(0, 3);
+            }
+
+            return columnId;
+        }
+
+        function getColumnDescription (column) {
+            let columnDescription = column.elementLabel;
+            if (column.aggregation) {
+                const aggregationDescription = getAggregationDescription(column.aggregation);
+                columnDescription += ' (' + aggregationDescription + ')';
+            }
+
+            return columnDescription;
+        };
 
         function getCountColumn (col) {
             return {
@@ -91,7 +146,6 @@
                 filterPrompt: false,
                 id: col.id + 'ptc',
                 layerID: col.layerID,
-                objectLabel: col.objectLabel + ' count'
             };
         }
     }